Basic saving and loading....sorta
Basic saving and loading system implemented.  A solar system can be
saved and when reloaded. The planets appear to come back exactly as they
were, however, their orbits are a bit out of whack.  Could be a small
tweak to fix, could be a massive overhaul.

In this iteration, saving is accomplished via local storage.  The
implementation is primitive, thus only 1 solar system can be saved per
web browser.