Minor upload/download tweaks
engine.loadData and engine.updateFromLoad
Cleaned up the functions for loading data by standardizing the number of
arugments rather than delving deeper into optional argument hell.  Also
put in a logical order to the parameters of updateFromLoad.

exportData
When exporting orbit data, in the event that the user does not supply a
filename in the text box, the file will download with the default
filename "OrbitData.json" but if a filename is provided, it will use
that.  No significant checking is done to ensure that the filenames are
valid, only checking to make sure there's more than simply whitespace.